Changed the source generator to use the hoist pattern
Instead of having the users source generator be the entry point we create the entrypoint. This way we can 100% make sure the assembly resolver loaded before the users types get loaded. The main problem was that if users source generator contained static fields of types contained within external assemblies this would explode because the assembly resolver had not be subscribed yet